# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: codexws HTTP API
  version: 3.0.0
  description: 'Current same-origin API for the one shared codexws backend process.

    This contract documents deployed paths; it is not the proposed /api/v1 design.

    '
servers:
- url: /
...
paths:
...
  /api/workspaces/register:
    post:
      tags:
      - workspaces
      operationId: registerWorkspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - path
              properties:
                id:
                  type: string
                name:
                  type: string
                path:
                  type: string
                create:
                  type: boolean
                  default: false
      responses:
        '200':
          description: Registered workspace and initial terminal.
          content:
            application/json:
              schema:
                type: object
                required:
                - ok
                - workspace
                properties:
                  ok:
                    type: boolean
                  workspace:
                    $ref: '#/components/schemas/Workspace'
        '400':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /api/workspaces/close:
    post:
      tags:
      - workspaces
      operationId: closeWorkspace
      requestBody:
        $ref: '#/components/requestBodies/WorkspaceSelector'
      responses:
        '200':
          description: Workspace terminals are closed and the registration is hidden
            or removed.
          content:
            application/json:
              schema:
                type: object
                required:
                - ok
                - workspace_id
                - workspaces
                properties:
                  ok:
                    type: boolean
                  workspace_id:
                    type: string
                  removed:
                    type: boolean
                  workspaces:
                    type: array
                    items:
                      $ref: '#/components/schemas/Workspace'
        '400':
          $ref: '#/components/responses/Error'
  /api/workspaces/reorder:
    post:
      tags:
      - workspaces
      operationId: reorderWorkspaces
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - order
              properties:
                order:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Updated workspace order.
          content:
            application/json:
              schema:
                type: object
                required:
                - ok
                - order
                - workspaces
                properties:
                  ok:
                    type: boolean
                  order:
                    type: array
                    items:
                      type: string
                  workspaces:
                    type: array
                    items:
                      $ref: '#/components/schemas/Workspace'
        '400':
          $ref: '#/components/responses/Error'
  /api/workspace-agent/restart:
    post:
      tags:
      - workspaces
      operationId: restartWorkspaceAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - workspace_id
              properties:
                workspace_id:
                  type: string
                thread_id:
                  type: string
      responses:
        '200':
          description: Workspace agent restart was requested.
          content:
            application/json:
              schema:
                type: object
                required:
                - ok
                - workspace_id
                - agent_command
                properties:
                  ok:
                    type: boolean
                  workspace_id:
                    type: string
                  agent_command:
                    type: string
        '400':
          $ref: '#/components/responses/Error'
  /api/workspace-terminals:
    get:
      tags:
      - terminals
      operationId: listWorkspaceTerminals
      parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      responses:
        '200':
          description: Runtime terminal summaries for one workspace.
          content:
            application/json:
              schema:
                type: object
                required:
                - workspace
                - terminals
                properties:
                  workspace:
                    $ref: '#/components/schemas/Workspace'
                  terminals:
                    type: array
                    items:
                      $ref: '#/components/schemas/Terminal'
        '400':
          $ref: '#/components/responses/Error'
  /api/workspace-terminals/create:
    post:
      tags:
      - terminals
      operationId: createTerminal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTerminalCommand'
      responses:
        '200':
          description: Created terminal and updated workspace summary.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminalCommandResult'
        '400':
          $ref: '#/components/responses/Error'
  /api/workspace-terminals/resume:
    post:
      tags:
      - terminals
      - conversations
      operationId: resumeConversationTerminal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/CreateTerminalCommand'
              - type: object
                required:
                - thread_id
                properties:
                  thread_id:
                    type: string
      responses:
        '200':
          description: Reused or created a terminal and queued at most one resume
            command.
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/TerminalCommandResult'
                - type: object
                  properties:
                    created:
                      type: boolean
                    reused:
                      type: boolean
                    resume_started:
                      type: boolean
        '400':
          $ref: '#/components/responses/Error'
  /api/workspace-terminals/close:
    post:
      tags:
      - terminals
      operationId: closeTerminal
      requestBody:
        $ref: '#/components/requestBodies/TerminalSelector'
      responses:
        '200':
          description: Terminal process was closed.
          content:
            application/json:
              schema:
                type: object
                required:
                - ok
                - workspace_id
                - terminal_id
                properties:
                  ok:
                    type: boolean
                  workspace_id:
                    type: string
                  terminal_id:
                    type: string
                  workspace:
                    $ref: '#/components/schemas/Workspace'
        '400':
          $ref: '#/components/responses/Error'
  /api/workspace-terminals/rename:
    post:
      tags:
      - terminals
      operationId: renameTerminal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/TerminalSelector'
              - type: object
                required:
                - name
                properties:
                  name:
                    type: string
                    maxLength: 80
      responses:
        '200':
          description: Renamed terminal summary.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminalCommandResult'
        '400':
          $ref: '#/components/responses/Error'
  /api/terminal-buffer:
    get:
      tags:
      - terminals
      operationId: getTerminalBuffer
      parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - $ref: '#/components/parameters/TerminalId'
      - name: after
        in: query
        schema:
          type: integer
          minimum: 0
      - name: before
        in: query
        schema:
          type: integer
          minimum: 0
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
      responses:
        '200':
          description: Absolute-offset slice of terminal history.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminalBuffer'
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /api/terminal-screen:
    get:
      tags:
      - terminals
      operationId: getTerminalScreen
      parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - $ref: '#/components/parameters/TerminalId'
      - name: stream_id
        in: query
        schema:
          type: string
      - name: after_seq
        in: query
        schema:
          type: integer
      - name: reason
        in: query
        schema:
          type: string
      responses:
        '200':
          description: Authoritative terminal frame used for attach or recovery.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminalFrame'
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
...
  /api/workspaces/pin:
    post:
      summary: Persist workspace pin order
      responses:
        '200':
          description: Success
        '400':
          description: Invalid request
        '409':
          description: Unavailable or conflicting state
  /api/workspace-terminals/agent:
    post:
      summary: Start the optional CLI in a dedicated PTY
      responses:
        '200':
          description: Success
        '400':
          description: Invalid request
        '409':
          description: Unavailable or conflicting state
  /api/workspace-terminals/interrupt:
    post:
      summary: Send Ctrl-C; completion must be checked in terminal
      responses:
        '200':
          description: Success
        '400':
          description: Invalid request
        '409':
          description: Unavailable or conflicting state
...
components:
  parameters:
    WorkspaceId:
      name: workspace_id
      in: query
      required: true
      schema:
        type: string
        minLength: 1
    TerminalId:
      name: terminal_id
      in: query
      required: true
      schema:
        type: string
        minLength: 1
...
  requestBodies:
    WorkspaceSelector:
      required: true
      content:
        application/json:
          schema:
            type: object
            required:
            - workspace_id
            properties:
              workspace_id:
                type: string
                minLength: 1
    TerminalSelector:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TerminalSelector'
  responses:
    Error:
      description: Request failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required:
      - error
      properties:
        error:
          type: string
...
    Workspace:
      type: object
      required:
      - id
      - name
      - cwd
      additionalProperties: true
      properties:
        id:
          type: string
        name:
          type: string
        cwd:
          type: string
        source:
          type: string
        exists:
          type: boolean
        terminal:
          $ref: '#/components/schemas/Terminal'
        terminals:
          type: array
          items:
            $ref: '#/components/schemas/Terminal'
    Terminal:
      type: object
      required:
      - id
      - workspace_id
      additionalProperties: true
      properties:
        id:
          type: string
        workspace_id:
          type: string
        name:
          type: string
        cwd:
          type:
          - string
          - 'null'
        alive:
          type: boolean
        usable:
          type: boolean
        pid:
          type:
          - integer
          - 'null'
        thread_id:
          type:
          - string
          - 'null'
        thread_title:
          type:
          - string
          - 'null'
    TerminalSelector:
      type: object
      required:
      - workspace_id
      - terminal_id
      properties:
        workspace_id:
          type: string
          minLength: 1
        terminal_id:
          type: string
          minLength: 1
    CreateTerminalCommand:
      type: object
      required:
      - workspace_id
      properties:
        workspace_id:
          type: string
          minLength: 1
        terminal_id:
          type: string
        name:
          type: string
          maxLength: 80
        cols:
          type: integer
          minimum: 2
          maximum: 1000
        rows:
          type: integer
          minimum: 1
          maximum: 500
    TerminalCommandResult:
      type: object
      required:
      - ok
      - workspace
      - terminal
      properties:
        ok:
          type: boolean
        workspace:
          $ref: '#/components/schemas/Workspace'
        terminal:
          $ref: '#/components/schemas/Terminal'
    TerminalBuffer:
      type: object
      required:
      - workspace_id
      - terminal_id
      - buffer
      - start
      - end
      - history_start
      - history_end
      additionalProperties: true
      properties:
        workspace_id:
          type: string
        terminal_id:
          type: string
        buffer:
          type: string
        start:
          type: integer
        end:
          type: integer
        history_start:
          type: integer
        history_end:
          type: integer
        has_more:
          type: boolean
    TerminalFrame:
      type: object
      required:
      - workspace_id
      - terminal_id
      - stream_id
      - frame_seq
      additionalProperties: true
      properties:
        workspace_id:
          type: string
        terminal_id:
          type: string
        stream_id:
          type: string
        stream_epoch:
          type: integer
        frame_seq:
          type: integer
        frame_end:
          type: integer
        screen_snapshot:
          type: string
        screen_snapshot_delta_safe:
          type: boolean
...